Used MVP for refactoring
