+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,0 +1,30 @@
+import Foundation
+import UIKit
+
+final class MovieQuizPresenter {
+    
+    // переменная с индексом текущего вопроса, начальное значение 0
+    private var currentQuestionIndex = 0
+    //общее количество вопросов для квиза
+    let questionsAmount: Int = 10
+    
+    func isLastQuestion() -> Bool {
+            currentQuestionIndex == questionsAmount - 1
+        }
+        
+        func resetQuestionIndex() {
+            currentQuestionIndex = 0
+        }
+        
+        func switchToNextQuestion() {
+            currentQuestionIndex += 1
+        }
+    
+    func convert(model: QuizQuestion) -> QuizStepViewModel {
+        let questionStep = QuizStepViewModel(
+            image: UIImage(data: model.image) ?? UIImage(),
+            question: model.text,
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount)")
+        return questionStep
+    }
+}
